Enable caller to use a root folder path to untie scores
Suppose you have the following files in your tree
- `/A/package.json`
- `/A/B/package.json`
- `/A/B/C/package.json`
- `/A/C/D/package.json`
- `/A/REALLY_BIG_NAME/package.json`

and you are editing a file in `/A/B/C` directory. If you query for
`package`, you are going to get `/A/package.json` as first suggestion,
since the current logic to break ties prefers the suggestions with few
characters. We can do better if we break the tie considering the
distance between the files suggestion to the current directory. In this
case, the suggestion should be `/A/B/C/package.json`.